+++ server/src/mapobjects.cpp
@@ -791,7 +791,7 @@ inline cUObject* nextObjectInCircle( IteratorState* is )
 			const Coord &pos = object->pos();
 			SI16 dx = ( pos.x - state.x );
 			SI16 dy = ( pos.y - state.y );
-			if ( ( ( dx * dx ) + ( dy * dy ) ) <= state.radius2 )
+			if ( (unsigned int)( ( dx * dx ) + ( dy * dy ) ) <= state.radius2 )
 			{
 				return object;
 			}
